Finalmente

+++ library/include/Reception.h
@@ -2,13 +2,14 @@
 #define PROJEKTX_RECEPTION_H
 
 #include "RentsRepository.h"
+#include "ExceptionRent.h"
 
 class Reception {
 private:
     RentsRepository currentRentsRepository;
     RentsRepository archiveRentsRepository;
 public:
-    int getRentPrice(shared_ptr<Rent>);
+    double getRentPrice(shared_ptr<Rent>);
     void createRent(shared_ptr<Client>,shared_ptr<Room>, string);
     double endRent(shared_ptr<Rent>);
     shared_ptr<Rent> getRent(shared_ptr<Room>);
